organization of dom manipulation files
