+++ browser-search-regex/chrome/custom_lib/DOM_manipulator/Globals.ts
