+++ browser-search-regex/chrome/manifest.json
@@ -7,7 +7,11 @@
   "content_scripts": [
     {
       "matches": ["*://*/*", "https://*/"],
-      "js": ["./custom_lib/stylers/Styler.js"],
+      "js": [
+        "./custom_lib/stylers/Styler.js",
+        "./custom_lib/DOM_manipulator/DomPopup.js",
+        "./custom_lib/DOM_manipulator/Globals.js"
+      ],
       "css": []
     }
   ],
